Move some util functions into a new file.
This avoids a circular import when using util functions from
warpinator-send.
