+++ src/misc.py
@@ -1,0 +1,28 @@
+#!/usr/bin/python3
+
+import threading
+import traceback
+
+from gi.repository import GLib
+
+# Used as a decorator to run things in the background
+def _async(func):
+    def wrapper(*args, **kwargs):
+        thread = threading.Thread(target=func, args=args, kwargs=kwargs)
+        thread.daemon = True
+        thread.start()
+        return thread
+    return wrapper
+
+# Used as a decorator to run things in the main loop, from another thread
+def _idle(func):
+    def wrapper(*args, **kwargs):
+        GLib.idle_add(func, *args, **kwargs)
+    return wrapper
+
+def print_stack():
+    traceback.print_stack()
+
+def check_ml(fid):
+    on_ml = threading.current_thread() == threading.main_thread()
+    print("%s on mainloop: " % fid, on_ml)
